requirements.txt
Django==3.2.8
django-cors-headers==3.10.0
djangorestframework==3.12.4
djangorestframework-jwt==1.0.1
djangorestframework-simplejwt==5.0.0
PyJWT==2.3.0
PyMySQL==1.0.2
pytz==2021.3
sqlparse==0.4.2
mysqlclient
pymysql
markdown 
django-filter
django-cors-headers